Move AndroidX-related code in core-android into a separate package.
More such code is coming, and this module is getting a bit messy.
